VM location aware of Full-mesh Tenant-specific GRE Tunnel support over OF1.0 datapath.
This push also includes the following major changes :
1. OVSDB Port tagging on the internal integration bridge (br-int) on Tenant VM connected ports
2. Proper Container support with Flow Spec (internal Vlan) and Openflow ports as Container NodeConnectors
3. Moved most of the logic away from the Neutron Northbound Handlers to the Southbound Handler
4. More fixes to make the OVSDB bundles restartable without the need to force restart the Openstack components.
5. Moved the Southbound events to be handled in its own event thread.

+++ ovsdb/src/main/java/org/opendaylight/ovsdb/lib/table/Interface.java
@@ -1,5 +1,7 @@
 package org.opendaylight.ovsdb.lib.table;
 
+import java.math.BigInteger;
+
 import org.opendaylight.ovsdb.lib.notation.OvsDBMap;
 import org.opendaylight.ovsdb.lib.notation.OvsDBSet;
 import org.opendaylight.ovsdb.lib.table.internal.Table;
@@ -13,7 +15,7 @@ public class Interface extends Table<Interface> {
     private String name;
     private OvsDBMap<String, String> options;
     private String type;
-    private OvsDBSet<Integer> ofport;
+    private OvsDBSet<BigInteger> ofport;
     private OvsDBSet<String> mac;
     private OvsDBMap<String, Integer> statistics;
     private OvsDBMap<String, String> status;
@@ -44,11 +46,11 @@ public class Interface extends Table<Interface> {
         this.type = type;
     }
 
-    public OvsDBSet<Integer> getOfport() {
+    public OvsDBSet<BigInteger> getOfport() {
         return ofport;
     }
 
-    public void setOfport(OvsDBSet<Integer> ofport) {
+    public void setOfport(OvsDBSet<BigInteger> ofport) {
         this.ofport = ofport;
     }
